refactoring notification and relation

+++ heaptrip/src/test/java/com/heaptrip/service/account/community/InitCommunityTest.java
@@ -3,7 +3,8 @@ package com.heaptrip.service.account.community;
 import com.heaptrip.domain.entity.account.relation.Relation;
 import com.heaptrip.domain.repository.account.relation.RelationRepository;
 import com.heaptrip.domain.repository.solr.SolrAccountRepository;
-import com.heaptrip.domain.service.account.criteria.RelationCriteria;
+import com.heaptrip.domain.service.account.criteria.relation.AccountRelationCriteria;
+import com.heaptrip.domain.service.account.criteria.relation.RelationCriteria;
 import com.heaptrip.service.account.user.UserDataProvider;
 import org.apache.solr.client.solrj.SolrServerException;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,12 +13,10 @@ import org.springframework.test.context.testng.AbstractTestNGSpringContextTests;
 import org.testng.annotations.AfterTest;
 import org.testng.annotations.BeforeTest;
 
-import com.heaptrip.domain.entity.account.community.club.Club;
 import com.heaptrip.domain.service.account.community.CommunityService;
 
 import java.io.IOException;
 import java.util.List;
-import java.util.Locale;
 
 @ContextConfiguration("classpath*:META-INF/spring/test-context.xml")
 public class InitCommunityTest  extends AbstractTestNGSpringContextTests {
@@ -35,12 +34,13 @@ public class InitCommunityTest  extends AbstractTestNGSpringContextTests {
     public void init() throws Exception {
         this.springTestContextPrepareTestInstance();
         deleteAccounts();
+        deleteRelation();
     }
 
     @AfterTest
     public void afterTest() {
         deleteAccounts();
-        deleteRelations();
+        deleteRelation();
     }
 
     private void deleteAccounts() {
@@ -50,33 +50,20 @@ public class InitCommunityTest  extends AbstractTestNGSpringContextTests {
         communityService.hardRemove(CommunityDataProvider.DELETED_COMMUNITY_ID);
     }
 
-    private void deleteRelations() {
-        RelationCriteria relationCriteria = new RelationCriteria();
-        relationCriteria.setFromId(UserDataProvider.EMAIL_USER_ID);
+    private void deleteRelation() {
+        List<Relation> relations = relationRepository.findByAccountRelationCriteria(new AccountRelationCriteria(CommunityDataProvider.COMMUNITY_ID));
 
-        List<Relation> relations = relationRepository.findByCriteria(relationCriteria);
+        if (relations != null) {
+            for (Relation relation : relations) {
+                relationRepository.remove(relation.getId());
 
-        for (Relation relation : relations) {
-            relationRepository.remove(relation.getId());
-        }
-
-        for (Relation relation : relations) {
-            try {
-                solrAccountRepository.remove(relation.getFromId());
-            } catch (SolrServerException e) {
-                e.printStackTrace();
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        }
-
-        for (Relation relation : relations) {
-            try {
-                solrAccountRepository.remove(relation.getToId());
-            } catch (SolrServerException e) {
-                e.printStackTrace();
-            } catch (IOException e) {
-                e.printStackTrace();
+                try {
+                    solrAccountRepository.remove(relation.getFromId());
+                } catch (SolrServerException e) {
+                    e.printStackTrace();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
             }
         }
     }
